WADNR:
Boy, this untouchable person stuff is a real aggravation. This code:

        public static Person Person
        {
            get
            {
                return GetValueOrDefault(PersonKey,
                    () => Saml2ClaimsHelpers.GetOpenIDUserFromPrincipal(GetHttpContextUserThroughOwin(),
                        Person.GetAnonymousSitkaUser()));
            }
            set { SetValue(PersonKey, value); }
        }

Was surely a convenient hack at the time, but is a minor nightmare now. Once you have a security exception you are pinned down and can't move without re-touching Person and blowing a double exception.

This fork demonstrates the problem, and shows some fitful progress towards a hacky solution, but it's definitely nowhere near ready for prime time.